+++ src/Client.php
@@ -17,13 +17,13 @@ use Http\Client\HttpAsyncClient;
 use Http\Discovery\HttpAsyncClientDiscovery;
 use Http\Discovery\Psr17FactoryDiscovery;
 use Http\Message\Authentication\Bearer;
-use Paddle\SDK\Entities\DateTime;
 use Paddle\SDK\Logger\Formatter;
 use Paddle\SDK\Resources\Addresses\AddressesClient;
 use Paddle\SDK\Resources\Adjustments\AdjustmentsClient;
 use Paddle\SDK\Resources\Businesses\BusinessesClient;
 use Paddle\SDK\Resources\CustomerPortalSessions\CustomerPortalSessionsClient;
 use Paddle\SDK\Resources\Customers\CustomersClient;
+use Paddle\SDK\Resources\DiscountGroups\DiscountGroupsClient;
 use Paddle\SDK\Resources\Discounts\DiscountsClient;
 use Paddle\SDK\Resources\Events\EventsClient;
 use Paddle\SDK\Resources\EventTypes\EventTypesClient;
@@ -48,14 +48,6 @@ use Psr\Http\Message\UriFactoryInterface;
 use Psr\Http\Message\UriInterface;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
-use Symfony\Component\Serializer\Encoder\JsonEncoder;
-use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter;
-use Symfony\Component\Serializer\Normalizer\AbstractObjectNormalizer;
-use Symfony\Component\Serializer\Normalizer\BackedEnumNormalizer;
-use Symfony\Component\Serializer\Normalizer\DateTimeNormalizer;
-use Symfony\Component\Serializer\Normalizer\JsonSerializableNormalizer;
-use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
-use Symfony\Component\Serializer\Serializer;
 use Symfony\Component\Uid\Ulid;
 
 class Client
@@ -74,6 +66,7 @@ class Client
     public readonly AddressesClient $addresses;
     public readonly BusinessesClient $businesses;
     public readonly DiscountsClient $discounts;
+    public readonly DiscountGroupsClient $discountGroups;
     public readonly SubscriptionsClient $subscriptions;
     public readonly EventTypesClient $eventTypes;
     public readonly EventsClient $events;
@@ -123,6 +116,7 @@ class Client
         $this->customerPortalSessions = new CustomerPortalSessionsClient($this);
         $this->businesses = new BusinessesClient($this);
         $this->discounts = new DiscountsClient($this);
+        $this->discountGroups = new DiscountGroupsClient($this);
         $this->subscriptions = new SubscriptionsClient($this);
         $this->eventTypes = new EventTypesClient($this);
         $this->events = new EventsClient($this);
@@ -192,20 +186,8 @@ class Client
 
         $request = $this->requestFactory->createRequest($method, $uri);
 
-        $serializer = new Serializer(
-            [
-                new BackedEnumNormalizer(),
-                new DateTimeNormalizer([DateTimeNormalizer::FORMAT_KEY => DateTime::PADDLE_RFC3339]),
-                new JsonSerializableNormalizer(),
-                new ObjectNormalizer(nameConverter: new CamelCaseToSnakeCaseNameConverter()),
-            ],
-            [new JsonEncoder()],
-        );
-
         if ($payload !== null) {
-            $body = $serializer->serialize($payload, 'json', [
-                AbstractObjectNormalizer::PRESERVE_EMPTY_OBJECTS => true,
-            ]);
+            $body = JsonEncoder::default()->encode($payload);
 
             $request = $request->withBody(
                 // Satisfies empty body requests.