+++ src/Resources/Simulations/Operations/Config/Subscription/Renewal/SubscriptionRenewalEntitiesCreate.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Paddle\SDK\Resources\Simulations\Operations\Config\Subscription\Renewal;
+
+use Paddle\SDK\FiltersUndefined;
+use Paddle\SDK\Undefined;
+
+class SubscriptionRenewalEntitiesCreate implements \JsonSerializable
+{
+    use FiltersUndefined;
+
+    public function __construct(
+        public readonly string|Undefined|null $subscriptionId = new Undefined(),
+    ) {
+    }
+
+    public function jsonSerialize(): \stdClass
+    {
+        return (object) $this->filterUndefined([
+            'subscription_id' => $this->subscriptionId,
+        ]);
+    }
+}